Align pattern_attributes to 4-bytes boundaries This was reported as being an issue on failed render tests on AMD graphics cards

+++ src/data/bucket/pattern_attributes.js
@@ -5,8 +5,8 @@ const patternAttributes = createLayout([
     // [tl.x, tl.y, br.x, br.y]
     {name: 'a_pattern_from', components: 4, type: 'Uint16'},
     {name: 'a_pattern_to', components: 4, type: 'Uint16'},
-    {name: 'a_pixel_ratio_from', components: 1, type: 'Uint8'},
-    {name: 'a_pixel_ratio_to', components: 1, type: 'Uint8'},
+    {name: 'a_pixel_ratio_from', components: 1, type: 'Uint16'},
+    {name: 'a_pixel_ratio_to', components: 1, type: 'Uint16'},
 ]);
 
 export default patternAttributes;
